Split CompForm into two separate entry points
